feat: Display loading indicator for Quran data
Introduces a loading indicator in the ResourcesScreen while the Quran data is being prepopulated. This provides a better user experience by showing that the app is working on fetching the necessary data.

The `SunnahAssistantViewModel` now exposes a `prepopulateQuranDataCompletionStatus` SharedFlow to indicate the completion status of the Quran data prepopulation. The `ResourcesFragment` observes this status and passes it to the `ResourcesScreen`. The `ResourcesScreen` conditionally displays either the Quran data or a `CircularProgressIndicator` based on this status.